fixed includes added copyright
