+++ examples/example3.cpp
@@ -65,17 +65,17 @@ int example3()
                       "\n"
                       "{{# company}}"
                       "Company: {{name}}\n"
-	   "Invoice again: {{invoice_number}}\n"
-	   "{{#invoice_number}}\n"
-	   "has invoice\n"
-	   "{{/invoice_number}}\n"
-	   "{{#paied}}\n"
-	   "has paied\n"
-	   "{{/paied}}\n"
-	   "{{^paied}}\n"
-	   "has not paied\n"
-	   "{{/paied}}\n"
-	   "         {{street}}\n"
+					   "Invoice again: {{invoice_number}}\n"
+					   "{{#invoice_number}}\n"
+					   "has invoice\n"
+					   "{{/invoice_number}}\n"
+					   "{{#paied}}\n"
+					   "has paied\n"
+					   "{{/paied}}\n"
+					   "{{^paied}}\n"
+					   "has not paied\n"
+					   "{{/paied}}\n"
+					   "         {{street}}\n"
                       "         {{city}}, {{state}}  {{zip}}\n"
                       "{{/ company}}"
                       "------------------------------------------------\n"
