+++ include/boost/boostache/backend/detail/django_compiler.hpp
@@ -2,6 +2,7 @@
  *  \file detail/django_compiler.hpp
  *
  *  Copyright 2014 Michael Caisse : ciere.com
+ *  Copyright 2017 Tobias Loew : die-loews.de
  *
  *  Distributed under the Boost Software License, Version 1.0. (See accompanying
  *  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
